Restructure
Move everything frontend and register / login related to an express directory.

Do some refactoring to have CognitoClient throw exceptions for controllers to catch rather than the client catching an error and then throwing.

Make it possible to register with Cognito and verify email and telephone numbers - logic is still required to handle journeys that stop and are restarted - mostly this will be checking the state of the user and attributes using AdminGetUser and determining which path to take.  I expect the controllers to undergo further refactoring as these paths become apparent.

Add a login screen so registered users can log in.

For the most part this commit is to enable other people to collaborate but does cover the roughest cut of the registration flow and an equally rough version of the login flow.