Alternative CognitoClient
Provide an alternative implementation of the CognitoClient to enable development and testing without needing an AWS account / tokens
Alter `app.ts` to import the correct CognitoClient based on an environment variable - default to the actual implementation.
Update `package.json` so that the dev script uses StubCognitoClient
Provide a config for the StubCognitoClient so that calls to `createUser` can throw a `UsernameExistsException` if the email address inuse@foo.gov.uk is entered.
Provide an alternative format for the stub config to improve the implementation.  This is noise in this commit but it's a WIP.