Add nodemon back as it's needed for `npm run dev`
Put loading of allowed email domains into a module so it can be included and the file is only loaded once.

Move the rfc822 code into the lib folder and structure as a module. Rename function. Modify README to reflect that we've made changes.

Use import instead of require in emailValidator.ts to be consistent.  Import allowedEmalDomains and change domain checking function accordingly.  Remove try/catch blocks where error is unhandled. Set the email address in the session to the validated value before moving to controller so that if the email is trimmed, we end up storing the trimmed value.

Edit tests in emailValidator.spec.ts to not use callbacks.  Add descriptive constants to use in tests and messages returned from tests.  Altered description of tests to better describe the behaviour we are testing. Add extra test for RFC822 stuff. Others may be added later.  Removed test for `errorResponse` as it essentially only tests that we pass in some values and that a hard coded view is rendered.  This behaviour will be better tested in feature tests.