DFA-732 Use lambda to store details of user
Use LambdaFacade to do call to lambda from frontend.

Facade simplifies the creation of the payload and abstracts away the use of an Axios client to make the calls.  Much of the code that lived in the controller previously moves to the facade.

Add types to make working with TableItems and users easier.

Add cognitoUser to session and selfServiceUser to session - the cognitoUser contains stuff needed for sign-up / sign-in, the selfServiceUser is what gets stored.  Better names and arrangement will hopefully become clear in the future.