DFA-728 Store Client Details in DynamoDB
The data pushed in and out of the lambdas can be refactored neatly and the step function transformations can be used so that only the necessary payloads are passed in.  For the sake of speed this hasn't been done but will definitely need to be refactored at a later date.

Modify register-client.ts lambda to use correct parts of payload when calling Auth API Endpoint. Pass on enough information to put-client for save to DDB to be successful
Modify put-client to take the correct things from input and save to DDB.  Return enough information for next put-service-client.ts lambda to be successful
Modufy put-service-client lambda to store the correct information in DDB.
Modify template.yaml - have putClientFunction use nodejs16.x so that randomUuid is available in the runtime
Modify manage-account controller to log the output of generateClient