DFA-957 Backend for updating client details
Lambdas and step functions for updating a client's details -

* `do-update-client.ts` as entry point to invoked step function
* `update-client-in-registry.ts` to do what it says on the tin
* `update-service-client.ts` to update our database

* `updateClient.json`

`DynamoClient.ts` updated with code to generate udpate statements.  This is a chore because DDB has a lot of reserved keywords like `data` which happens to be one of our attributes.  I also had problems getting the dynamo-utils to marshal an array so there's some code to make that work. This is not very elegant but we are have a deadline to meet so good enough is good enough for the time being.  We use for-i loops becase we need orders to definitely be preserved and there's a consensus that "they should be" but nobody really trusts it.

Update a bunch of views and controllers to use the path parameters we need for the backend calls to work.  Implement two of the updates with little or no validation to ensure that simple strings and arrays can be pushed into the database.

Delete a bunch of `console.log` lines.

Run IntelliJ reformat code on everything.