Role Resource for sending SMS
Add and use an IAM role that permits Cognito to send SMS messages.

The role and the SmsConfiguration for the Cognito User Pool share an ExternalId.  This is currently hard coded but could be replaced with a parameter.  I'm not sure whether it's needed as it's designed for third parties: https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_create_for-user_externalid.html