Create Middleware
Create some middleware to do stuff with the public key.  This moves the confusing logic out of the controller (which confusingly still lives in testing routes) and into a middleware function that has a fairly self-documenting name.

Put the Public Key function in a file in the lib directory.  This might be overkill but I think it makes unit tests that haven't been written a little easier to deal with that having to test the middleware.  Maybe.
Insert the middleware function into the route for updating the Public Key.