DFA-2024 Make logged in status available to views
Menus will change based on whether a user is logged in or not.  Take this status from the session and make it available to request-response cycles with a default middleware.

Stub in logic for OTP processing - this is the point at which we determine whether a user has logged in successfully. set `req.session.isSignedIn` to true until OTP code is implemented.
Create `setSignedInStatus.ts` middleware.  Set `res.locals.isSignedIn` to true if session exists and `req.session.isSignedIn` is true, otherwise set value to false.
Add a sign-out function to the sign-in controller `¯\_(ツ)_/¯`
Use above controller in sign-in routes - `/account/sign-out`
Point the sign-out link in the layout-dashboard.njk to the sign-out route.
Insert default middleware into middleware chain in app.ts.  Add isSignedIn to session definition in app.ts