Add some useful error handling to `changePassword` controller method - handle supplied password being wrong and too many attempts exception.
The error message for too many attempts does need updating and is just a place holder at the moment until some proper content can be designed.

Insert CheckAuthorisation to some routes to ensure we are logged in with a session before trying do things like change password.  This middleware is incredibly basic and should deal with refreshing Cognito tokens.  Applying it in all the right places should be a separate story in the future.