DFA-2010 User attempts to use a common password
Add `src/middleware/notOnCommonPasswordListValidator/` which contains an index file to export the part that matters, a class to encapsulate the functionality, a file to export a single instance (the part that matters) so we don't load 100,000 lines of text every time someone changes a password, some actual validation middleware which uses the singleton.

In a novel twist we include unit tests here which test the class and test a fairly minimal set of cases - a password on the list and one that's not on the list.

We're also committing file with the 100,000 passwords in it.