Implement feature tests.
Rename `features/support/steps/sign-in-steps.ts` to `features/support/steps/sign-up-steps.ts`
Add some candy coated steps to sign-in-steps to wrap input and submission of specific cases.
Create a generic version of the above steps in `shared-steps.ts` so we don't always need precisely named steps - it's more useful for finding out what the precisely named steps should be.
Create a function that gets called by the above two lots of steps.
Edit a bunch of steps so we 'submit' an email address instead of 'registering' it because I changed the wording of the shared step.
Create `sign-up.feature` - this only covers a successful sign up and one that crashes to a halt when the user tries a password on the list.  We can argue about how they should actually be structured at a later date - there's an argument for all the password related stuff being in one place etc.