BAU: Remove duplication in top nav and layouts
Refactor top nav block to not repeat code blocks. Use the `isSignedIn` variable to show or hide appropriate headers.

Use a template variable to show or hide the top navigation bar. The default is true and can be set to false in child templates of `layout.njk`.

Remove the `layout-no-top-nav.njk` template since it is identical to the main `layout.njk` template with the `showTopNav` variable set to `false`.