DFA-2048: Don't hard code paths to npm modules
npm modules are now installed in the root of the repo. Make the codebase more robust by using npm methods to derive the location of specific packages rather than hard coding a path.

In npm scripts: Use the `npm ls -p <package>` to get the path.

In Typescript code: Use the `require.resolve("<package>/package.json")` method to get the path to the root of the module.
See https://stackoverflow.com/questions/10111163/how-can-i-get-the-path-of-a-module-i-have-loaded-via-require-that-is-not-mine

In Sass: include the path to the npm module when compiling sass and use paths from the module's root in Sass files.