BAU: Common linting config
Add ESLint and Prettier config files in the root of the repo so the same rules are applied to all projects.
Reference: https://typescript-eslint.io/docs/linting/typed-linting/monorepos/#one-tsconfigjson-per-package-and-an-optional-one-in-the-root

Use the ESLint Typescript and Prettier plugin, and apply the recommended rules.

Use the ESLine projects feature to specify all projects that should have linting rules applied.
The `tsconfig.eslint.json` file in the root of the repo specifies additional files to lint not included by other tsconfig.json files in the repo.

The rules are currently not enforced, though we may decide to do so in the future. For now, it is recommended to set a keyboard shortcut in the IDE of choice to apply the eslint/prettier suggestions or to configure the IDE to apply them automatically when saving files.