DFA-2129 Abstract away backing services
Abstracting backing services and shifting stuff out of the session

Rely on returned Authentication results for Cognito related stuff and get rid of 'CognitoUser' from session.  Create a class with static method to extract relevant claims from AuthenticationResults

Move other functionality into S4.

Change mobile phone number validation.  Stop throwing errors and return values.  Prepare value for cognito in controller instead of validation middleware.

Overhaul StubCognitoClient to be more dynamic about the tokens that get returned.

Render the entered mobile number instead of what Cognito got.

Use RefreshToken to refresh the tokens in `checkAuthorisation` - this makes extracting things from tokens more likely to be up to date.

Add further methods to S4

Lay some foundations for Stubs to update values and keep some state.

Fix linting in tests because apostrophes are evil.  Apparently.

Add error messages so tests pass again - we're probably going to stop throwing errors anyway so a light touch is all we need at this point.

Change phone number in acceptance test because it displays what the user entered now.

Put calls to lambda facade and cognito client behind a single client.

Mostly this is just shadowing methods but there's some stuff that simplifies the arguments passed in.

Try to get as much information as possible from the access tokens and not from the session.

Remove redundant parameter from a lambda - it seemed at first that the same method could be used for updating the database but that's no longer true so an undefined id no longer makes sense.

Update StubLambdaFacade so that changes to the user are reflected in subsequent calls to retrieve the user.