Make Service a domain object
Update session in app.ts
Add to types to distinguish between what comes back from the lambdas and what we want to use.  This can be further refactored inside the lambdas in future.
Update lambdas to receive the domain service
Add Utils to change between dynamo / lambda output and the domain objects
Unrelated update to template that is rendered in processLoginOtpMobile