DFA-2048: Add a workflow to build and push frontend image to ECR
The workflow builds and pushes frontend images for preview deployments.

Each image is versioned with the git SHA and additionally has the current branch tag. Since the repository is immutable, only one image per branch is allowed. The previous version for a branch, if it exists, is deleted before pushing a newer one.

Images are built and pushed when a branch is deployed manually (for now).