DFA-2244: Pass correct parameters for the client details page
Add a return type for the lambdas calling the database API. For now, we're using the AWS SKD types but in the future we should switch to using our own types shared by the API and the frontend (the 'contract)'.

Temporarily Move the "@aws-sdk/client-dynamodb" package to the main `package.json` file as it's now used in multiple projects.