SSE-2385 Correctly render public key data
Public key data was not being saved in DDB because the wrong ID was being used for the client.  Also, the public key provided by the user and not the public key suitable for Auth was being passed through so even if the ID was correct, Auth would have rejected it.