SSE-2365 Make Change service name work on all pages
The Team members route did not include the parameters for `serviceId`, `selfServiceClientId` or `clientId` so it could not render the Change service link properly and couldn't propagate those IDs to the pages referenced in the side bar.

Change the route to include the parameters and make sure they are added to the rendering code so they can be added to URLs.

Change `views/service-details/layout-service-details.njk` so missing IDs are added to the team-members URL as parameters.