SSE-2386 Public Key not displayed in text area
Public Key data should be displayed in the text area used for entering public keys and it is not currently even though it's in the Figma and you never know when you'll need to manually edit the public key you've already successfully stored.

Use correctly sanitised version of public key so comparisons work and the default isn't displayed.

Pass value to template.

Add test.