SSE-2336: Fix timeout issue in private beta acceptance tests when opening a new tab
`Page.waitForTimeout()` is obsolete and should not be used. It is not recommended to wait for a set number of seconds but instead use a waiter method to wait exactly as long it takes for the desired conditions to be met.
Puppeteer docs: https://pptr.dev/api/puppeteer.page.waitfortimeout

Replace `waitForTimeout` with `waitForTarget` to wait until the new tab is opened by the current page. When the new tab is open, load the page contents and make it the current page for the text context so that later steps can reference it.

Split the step into smaller steps and add a check to verify the new page's URL and contents.

Remove the empty step about saving the user's data - this is not in scope of the acceptance tests and we don't want to leak the internal email address to GitHub.

Change the service name returned from the stub lambda facade to match what the tests are expecting.