BAU: Simplify button clicking in acceptance tests
The Design System renders the Button component either as an HTML `button` or `a` tag, depending on whether it's used a submit button in a form. Therefore, two methods are needed to find a regular button or a `button-link`. Simplify the helper methods to take additional params as needed.

Simplify the button clicking methods to automatically handle whether the button is a button or a link, to make it easier to use in features.

Use pathnames when checking the current page path to avoid having to run regex on URLs.

Change some step names for consistency.