SSE-2491 Fix Public Key Bug
Because we just manipulated the input to return and assumed that the crypto lib would refuse to convert input with junk after the key there was a bug.

We would append anythng on a new line after `-----END PUBLIC KEY-----` to the key we rendered and stored.

This changes that so that we export the converted key to PEM format and then strip that so it works for Auth.

Wrap stripping that makes it auth compliant in a function.