SSE-2418: Organise lambdas
Group the lambdas by categories and rename `dynamodb-api` to just `api` since it encapsulates a wider functionality.

Ideally, we'd have a stack per category to keep the templates and stacks small, and to make the deployments more modular, however, that would mean having to set up multiple secure pipelines so we need to balance the number of stacks vs the number of pipelines we want to maintain. We might want to break up the lambdas into multiple stacks/deployments in the future.