SSE-2496: Add a deployment bucket and additional permissions to the GitHub Actions role
More permissions are needed to deploy the API lambdas.
A SAM deployment bucket needs to be set up in the account before deployments can be done.

Add a check for the frontend infra stack and the OIDC provider, as the template will need to be used in different envs with different requirements.

Ignore a bunch of checkov warnings for now - we will deal with them later.