SSE-2496: Prefix step functions log group name to avoid CloudWatch log policy size error
There is one global CloudWatch logs resource policy used by CloudFormation (viewed by running `aws logs describe-resource-policies`) to grant the log delivery service permissions to send logs to log groups. When a state machine is created with logging enabled, CloudFormation checks if the policy grants the permission for the log group specified in the template.

If CloudWatch detects that the log delivery service doesn't have a permission to send logs to the log group being created, it will try to update the policy to grant the required access by adding a new statement. In effect, it adds a new statement for each unique log group name. This means that if enough log groups are created, the policy will exceed the allowed limit and all deployments will fail from that point on.

To prevent this from happening, prefix the state machines log group name and add a custom CloudWatch logs resource policy with a wildcard to allow sending logs to all subgroups.

References:
[Amazon CloudWatch Logs resource policy size restrictions](https://docs.aws.amazon.com/step-functions/latest/dg/bp-cwl.html)
[IAM Policies for logging to CloudWatch Logs](https://docs.aws.amazon.com/step-functions/latest/dg/cw-logs.html#cloudwatch-iam-policy)
[Bug with LoggingConfiguration | AWS re:Post](https://repost.aws/questions/QURc2glxBETSe3Q6Y0UwcpQg/bug-with-logging-configuration)