SSE-2515 Grafana Stacks in development
Scripts to capture the commands used to create Grafana stacks in development.  Both scripts run a shared version of the script with the correct parameter for environment.

The key-rotator script will check whether the secret exists (a pre-requisite for running the script) and will either use a provided environment variable for the value or take from a prompt.