Bugfix Post logout redirect URIs
Post logout redirect URIs became optional and as a result an empty database will not supply a value for them meaning clients can't be displayed because the frontend tries to do `.split(" ")` on an `undefined` value.

This ensures that if `client.post_logout_redirect_uris` are not returned, an empty array is returned instead.