SSE-2165: Mount routers on specific paths
This helps to modularise the app and use composition to combine smaller apps into the self-service system.
We also get more control over the paths and will be able to use this mechanism to mount stubs and testing paths as needed in non-prod environments.

Remove the mount path from routes. Routers specify paths relative to the path they've been mounted on.
Make redirects in routers relative (without the leading /).
Mount the testing routes on a path as well to avoid cross-contamination with the main app (some pages will need to be moved).