SSE-2575: Add a utility function to get optional values
The function takes a value or an object and a set of property names, and returns the values if they are present or throws an exception if values are undefined.

This simplifies checking for nulls at runtime and can be used where an optional value is not expected.

The function can be extended in the future to include a custom handler or a custom error message.

Register the function on `globalThis` and import the module in `app.ts` so it is available globally and doesn't need to be explicitly imported.