SSE-2575: Add context to requests
Extend the express request interface to include an object with optional pre-defined properties.
Extend the express application to initialise the context object with an empty value.

This works the same way as the `req.session` object but is intended to store information relevant to a single request-response cycle rather than the entire user's session.

This is useful when we want to mount an app on a sub-path that includes a parameter, since `req.params` won't be available to the child app. The parent app must set the params it wants to pass to the child app.

Express docs recommend setting properties directly on the request object but since we use TypeScript we need to pre-define the object and the properties it will hold to avoid errors and get typechecking.

Express reference:
https://expressjs.com/en/4x/api.html#router.param