BAU: Move acceptance tests out the express directory
The acceptance tests need to be run against the DOM TypeScript library in order to receive the correct types and checks for web page elements. Recent changes to Puppeteer broke some of the checks as some types and interfaces were changed, and the full DOM library is needed to run the tests with the latest version.

Move the tests to the top-level 'features' directory and set it up as a new project in the monorepo. Add config for npm, cucumber and TypeScript, and set the library to DOM.

Adjust scripts and workflows to use the new locations.
Add a timeout to the acceptance tests job.

Address the highlighted issues with the tests:
  - Handle possibly null values
  - Check the element type when running XPath queries to satisfy new Puppeteer interfaces
  - Increase the timeout when clicking on links to external sites that might take longer to load