BAU: Use the nullish coalescing operator
Use the nullish coalescing operator instead of the boolean OR when dealing with null or undefined values to provide a default value. The boolean || operator should only be used in boolean logic to evaluate an expression to true or false. Use the ?? operator to provide a value when a variable is null or undefined.