SSE-2685: Update the script to empty buckets for deletion
The script is now smarter and handles large buckets and can run in a loop whilst the delete job runs. The script runs in the background, keeps the buckets empty and automatically stops when all the buckets have been deleted.

The script now lists all stacks if no params have been passed and takes a param to select buckets for deletion.