SSE-2672: Turn on Cognito MFA and restore SMS config
Even though we use a custom lambda to send the MFA code, we still need to leave MFA enabled in the User Pool.
Enabling the option requires providing SMS config, even it's not used.

Throw an error if the response from the initiate auth command returns null values, as we never expect them to be missing.