SSE-2638: Make the backend API private by default
When the frontend runs in the VPC, it can only communicate with the backend API Gateway via a VPC endpoint, meaning the endpoint must be set to private and the VPC subnets needs to be provided.

When running the frontend locally, we need the API to be public, so add a param to enable that and not set the private endpoint config.

Note that the API URL is different when executed through a VPC endpoint, hence the conditional on the output value.

Use a REST API as it's the only one that supports private endpoints.