BAU: Adds sign-in controller tests
I have tested most of the controllers in this file, however I have chosen to leave some of the simple controllers alone. In the globalSignOut tests there is a slightly odd mock implementation for the  function. This is to allow the anonymous call back function to be executed when req.destroy is called which is how the actual code functions