BAU: Adds AWS SDK mocking packages
The first of these is a general package AWS recommend for mocking SDK calls in the v3 SDK[^1]. The second is one that provides some extended test matchers for jest tests which come from same package above[^2]. The idea of these is to make mocking and matching SDK calls easier. In this case it will make asserting against cognito calls easier, considering the private send command method.

[^1]- https://aws.amazon.com/blogs/developer/mocking-modular-aws-sdk-for-javascript-v3-in-unit-tests/
[^2] - https://github.com/m-radzikowski/aws-sdk-client-mock\#jest-matchers