BAU: adds a setup.ts file to frontend jest config
Here we have this weird kind of globalSetup thing for the environment variables. I have a suspicion that due to us declaring a constant object for cognito environment variables in express/src/config/environment.ts  we cannot use the regular setup files as we usually do. Instead we need to use a globalSetup file to ensure we capture the environment variables when constant is declared in the express/src/config/environment.ts file. If we don't do this we end up with undefined environment variables which means the tests don't run. Anyway now the tests run, here are the docs for this jest feature https://jestjs.io/docs/configuration\#globalsetup-string this was really fiddly to debug