FIX: Add eslint ignore comment to explicit any
Our CI checks on PR have not picked up this any before, however when running the same checks on our merge queue this generates a warning and fails the CI checks. To get eslint to ignore this I have added an ignore comment. We should define the type more explicitly and remove the any in future, but for now to avoid unneccesary changes I'm adding the ignore comment