SSE-2815: Derive the public key from query param
When we render the template intially we get the public key from the query parameters. Currently when we enter an invalid public key, the "Current public key" section dissappers. This is because the `req.body.authCompliantPublicKey` field is undefined if getAuthCompliantKey throws an error. However when we have already set a public key this looks odd. I have instead used the query param to render the current public key, if one is set, like we do normally.