BAU: Removed Request Handler Typing on client controllers
The RequestHandler type expects a function signature that returns void. For our asynchronous controllers, these return a Promise<void>. Our sonarCloud picks up this issue and flags it as a bug. Whilst the request handler typing is useful, it is generating a lot of unneccesary issues in sonarcloud, so I have removed this and instead have typed the controllers more explicitly