BAU: Updates the error handling in update lambda
We were previously catching errors and returning a status code and error
message. This meant we passed through client registry errors which made
it harder to debug issues. I have re-implemented the throwing with some
additonal logging as we do not want to pass these error states through
to the next part of the step function