ATO-925: Adds session.save() to showGetEmailForm
The next controller in the user flow, showCheckEmailForm relies on the
session containing the email address. However we redirect when submitting
the email address, which may happen faster than our session updating,
leading to the user being redirected back to the enter-email page. The
docs for express-session say "There are some cases where it is useful to
call this method (`session.save()`), for example, redirects..."[1] so
adding this session.save call in to ensure the session is updated before
the redirect occurs.
[deploy]

[1]- https://expressjs.com/en/resources/middleware/session.html