BAU: Allow dev deployments from branch
We already have a pipeline in the dev account, yet this action
uses a seperate setup to deploy[1] to the dev account. This setup
is painful to use and often doesn't work.

So we're replacing this with a much simpler deploy dev workflow
dispatch. This is basically the deploy dev workflow that runs in
the merge queue, so will deploy to the 'main' dev stack. This is
fine as the merge group workflow will deploy the latest changes in
main before running acceptance tests.

[1] - https://github.com/govuk-one-login/onboarding-self-service-infra/blob/main/dev-gha-role/deployment-config.template.yml