BAU: Bump puppeteer and pin version
Bumps puppeteer to version 22.15.0 to resolve a vulnerability on
in tar-fs[1]. Have looked at the changelog and can see this doesn't
include any breaking changes. Also ran `npm update tar-fs 3.0.8` to
ensure this was fully updated

This change also pins puppeteer as it was not perviously pinned
which made this harder to debug locally. I think we should check
that packages are pinned more generally in this repo

[1]- https://github.com/govuk-one-login/onboarding-self-service-experience/security/dependabot/88